Fix house modal button

+++ Web/RavenAge.Web.ViewModels/City/CityViewModel.cs
@@ -25,13 +25,13 @@
 
         public int Artillery { get; set; }
 
-        public int Gold { get; set; } //// Premium
+        public decimal Gold { get; set; } //// Premium
 
-        public int Silver { get; set; }
+        public decimal Silver { get; set; }
 
-        public int Stone { get; set; }
+        public decimal Stone { get; set; }
 
-        public int Wood { get; set; }
+        public decimal Wood { get; set; }
 
         public DefenceWallViewModel DefenceWall { get; set; }
